Chore: Clean up the code
