+++ package.json
@@ -308,7 +308,7 @@
     "pofile": "1.0.11",
     "progress": "2.0.3",
     "react-textarea-autosize": "8.3.4",
-    "snapmaker-react-icon": "1.25.0",
+    "snapmaker-react-icon": "1.26.0",
     "style-loader": "0.21.0",
     "stylint": "1.5.9",
     "stylus": "0.54.8",
